friends combo box is changing
Whenever the  group combo box changes, the person combo box also changes
to its respective groups.

+++ group.cpp
@@ -31,7 +31,7 @@ Person* Group::getLatestPerson()
     return peopleRecords.fetchLatestRecord();
 }
 
-int Group::getPeopleCount()
+size_t Group::getPeopleCount()
 {
     return peopleRecords.getSize();
 }
